pyproject.toml
[project]
name = "project-school-profile"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dj-database-url>=3.0.1",
    "django-cotton>=2.3.0",
    "dotenv>=0.9.9",
    "psycopg>=3.2.12",
    "python-decouple>=3.8",
    "wagtail>=7.1.2",
]

[tool.pyright]
typeCheckingMode = "basic"
